+++ android/src/main/kotlin/com/coralogix/flutter/plugin/mappers/cx_rum/CustomMeasurementContextMapper.kt
@@ -6,15 +6,15 @@ import com.coralogix.flutter.plugin.mappers.IMapper
 class CustomMeasurementContextMapper : IMapper<CustomMeasurementContext?, Map<String, Any?>?> {
     override fun toMap(input: CustomMeasurementContext?): Map<String, Any?>? = input?.let {
         mapOf(
-            "name" to input.name,
-            "value" to input.value
+            "name" to it.name,
+            "value" to it.value
         )
     }
 
     override fun fromMap(input: Map<String, Any?>?): CustomMeasurementContext? = input?.let {
         CustomMeasurementContext(
-            name = input["name"] as? String ?: "",
-            value = input["value"] as? Long ?: 0
+            name = it["name"] as? String ?: "",
+            value = it["value"] as? Long ?: 0
         )
     }
 }