+++ android/src/main/kotlin/com/coralogix/flutter/plugin/mappers/cx_rum/EditableErrorContextMapper.kt
@@ -9,12 +9,13 @@ class EditableErrorContextMapper : IMapper<EditableErrorContext?, Map<String, An
 
     override fun toMap(input: EditableErrorContext?): Map<String, Any?>? = input?.let {
         mapOf(
-            "message" to input.message,
-            "type" to input.type,
-            "isCrash" to input.isCrash,
-            "stacktrace" to input.stacktrace?.map { frame ->
-                if (frame is EditableCoralogixAndroidStackFrame)
-                    frameMapper.toMap(frame)
+            "message" to it.message,
+            "type" to it.type,
+            "isCrash" to it.isCrash,
+            "stacktrace" to it.stacktrace?.mapNotNull { frame ->
+                (frame as? EditableCoralogixAndroidStackFrame)?.let { androidFrame ->
+                    frameMapper.toMap(androidFrame)
+                }
             }
         )
     }
@@ -22,10 +23,10 @@ class EditableErrorContextMapper : IMapper<EditableErrorContext?, Map<String, An
     @Suppress("UNCHECKED_CAST")
     override fun fromMap(input: Map<String, Any?>?): EditableErrorContext? = input?.let {
         EditableErrorContext(
-            message = input["message"] as? String ?: "",
-            type = input["type"] as? String ?: "",
-            isCrash = input["isCrash"] as? Boolean ?: false,
-            stacktrace = (input["stacktrace"] as? List<Map<String, Any?>>)?.map { frameMap ->
+            message = it["message"] as? String ?: "",
+            type = it["type"] as? String ?: "",
+            isCrash = it["isCrash"] as? Boolean ?: false,
+            stacktrace = (it["stacktrace"] as? List<Map<String, Any?>>)?.map { frameMap ->
                 frameMapper.fromMap(frameMap)
             }
         )