+++ android/src/main/kotlin/com/coralogix/flutter/plugin/mappers/cx_rum/InteractionContextMapper.kt
@@ -6,17 +6,17 @@ import com.coralogix.flutter.plugin.mappers.IMapper
 class InteractionContextMapper : IMapper<InteractionContext?, Map<String, Any?>?> {
     override fun toMap(input: InteractionContext?): Map<String, Any?>? = input?.let {
         mapOf(
-            "eventName" to input.eventName,
-            "targetId" to input.targetId,
-            "targetClassName" to input.targetClassName
+            "eventName" to it.eventName,
+            "targetId" to it.targetId,
+            "targetClassName" to it.targetClassName
         )
     }
 
     override fun fromMap(input: Map<String, Any?>?): InteractionContext? = input?.let {
         InteractionContext(
-            eventName = input["eventName"] as? String ?: "",
-            targetId = input["targetId"] as? String ?: "",
-            targetClassName = input["targetClassName"] as? String ?: ""
+            eventName = it["eventName"] as? String ?: "",
+            targetId = it["targetId"] as? String ?: "",
+            targetClassName = it["targetClassName"] as? String ?: ""
         )
     }
 }