+++ android/src/main/kotlin/com/coralogix/flutter/plugin/mappers/cx_rum/LifecycleContextMapper.kt
@@ -6,15 +6,15 @@ import com.coralogix.flutter.plugin.mappers.IMapper
 class LifecycleContextMapper : IMapper<LifecycleContext?, Map<String, Any?>?> {
     override fun toMap(input: LifecycleContext?): Map<String, Any?>? = input?.let {
         mapOf(
-            "event" to input.event,
-            "screenView" to input.screenView
+            "event" to it.event,
+            "screenView" to it.screenView
         )
     }
 
     override fun fromMap(input: Map<String, Any?>?): LifecycleContext? = input?.let {
         LifecycleContext(
-            event = input["event"] as? String ?: "",
-            screenView = input["screenView"] as? String ?: ""
+            event = it["event"] as? String ?: "",
+            screenView = it["screenView"] as? String ?: ""
         )
     }
 }