+++ android/src/main/kotlin/com/coralogix/flutter/plugin/mappers/cx_rum/LogContextMapper.kt
@@ -6,15 +6,15 @@ import com.coralogix.flutter.plugin.mappers.IMapper
 class LogContextMapper : IMapper<LogContext?, Map<String, Any?>?> {
     override fun toMap(input: LogContext?): Map<String, Any?>? = input?.let {
         mapOf(
-            "message" to input.message,
-            "data" to input.data
+            "message" to it.message,
+            "data" to it.data
         )
     }
 
     override fun fromMap(input: Map<String, Any?>?): LogContext? = input?.let {
         LogContext(
-            message = input["message"] as? String ?: "",
-            data = input["data"] as? String ?: ""
+            message = it["message"] as? String ?: "",
+            data = it["data"] as? String ?: ""
         )
     }
 }